Adapt cray- packages to spack package repo v2
