+++ site/spack_repo/alps/packages/cray_mpich/package.py
@@ -6,6 +6,8 @@
 import os
 import platform
 
+from spack_repo.builtin.build_systems.generic import Package
+
 import spack.compilers
 from spack.package import *
